+++ src/components/ConnectWidgets.tsx
@@ -1,11 +1,11 @@
 import {
   dispatchConnectLocationChangeEvent,
   ConnectPostMessageCallbackProps,
+  ConnectOAuthRequestedPayload,
 } from "@mxenabled/widget-post-message-definitions"
 
 import { Type, SsoUrlProps, ConnectWidgetConfigurationProps } from "../sso"
-import { LoadUrlInBrowserProps } from "./load_url_in_browser"
-import { makeDefaultConnectOnOAuthRequested } from "./oauth"
+import * as WebBrowser from "expo-web-browser"
 import { makeWidgetComponentWithDefaults } from "./make_component"
 import { useOAuthDeeplink, OAuthProps } from "./oauth"
 import { useWidgetRendererWithRef, StylingProps } from "./renderer"
@@ -13,7 +13,6 @@ import { useWidgetRendererWithRef, StylingProps } from "./renderer"
 export type ConnectWidgetProps = SsoUrlProps &
   StylingProps &
   OAuthProps &
-  LoadUrlInBrowserProps &
   ConnectPostMessageCallbackProps<string> &
   ConnectWidgetConfigurationProps &
   JSX.IntrinsicAttributes
@@ -24,17 +23,29 @@ export const ConnectVerificationWidget = makeWidgetComponentWithDefaults(Connect
 })
 
 export function ConnectWidget(props: ConnectWidgetProps) {
+  const onOAuthRequested = (payload: ConnectOAuthRequestedPayload) => {
+    const { url } = payload
+    WebBrowser.openAuthSessionAsync(url)
+
+    props.onOAuthRequested?.(payload)
+  }
+
   props = {
-    onOAuthRequested: makeDefaultConnectOnOAuthRequested(props),
+    onOAuthRequested,
     ...props,
   }
 
+  const modifiedProps = {
+    ...props,
+    onOAuthRequested,
+  }
+
   const [ref, elem] = useWidgetRendererWithRef(
-    { ...props, widgetType: Type.ConnectWidget },
+    { ...modifiedProps, widgetType: Type.ConnectWidget },
     dispatchConnectLocationChangeEvent,
   )
 
-  useOAuthDeeplink(ref, props)
+  useOAuthDeeplink(ref, modifiedProps)
 
   return elem
 }