+++ src/components/request_interceptor.test.ts
@@ -5,9 +5,15 @@ import {
   LoadPayload,
   dispatchConnectLocationChangeEvent as handler,
 } from "@mxenabled/widget-post-message-definitions"
+import { makeRequestInterceptor } from "./request_interceptor"
+import * as ExpoWebBrowser from "expo-web-browser"
 
-import { loadUrlInBrowser } from "../../src/components/load_url_in_browser"
-import { makeRequestInterceptor } from "../../src/components/request_interceptor"
+jest.mock("expo-web-browser", () => {
+  return {
+    openBrowserAsync: jest.fn().mockResolvedValue({ type: "success" }),
+  }
+})
+jest.mock("react-native")
 
 const makeNavigationEvent = (url: string) =>
   ({
@@ -22,7 +28,7 @@ const makeNavigationEvent = (url: string) =>
 
 describe("makeRequestInterceptor", () => {
   test("it returns a navigation event handler", () => {
-    const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+    const fn = makeRequestInterceptor("https://mx.com/", {
       onIntercept: (url) => handler(url, {}),
     })
     const req = makeNavigationEvent("https://mx.com/page2")
@@ -32,7 +38,7 @@ describe("makeRequestInterceptor", () => {
 
   describe("actions", () => {
     test("it bubbles the widget url back up to be loaded in the webview", () => {
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, {}),
       })
       const req = makeNavigationEvent("https://mx.com/")
@@ -41,7 +47,7 @@ describe("makeRequestInterceptor", () => {
     })
 
     test("it intercepts other urls", () => {
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, {}),
       })
       const req = makeNavigationEvent("https://mx.com/page2")
@@ -50,53 +56,27 @@ describe("makeRequestInterceptor", () => {
     })
 
     test("it intercepts post message urls", () => {
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, {}),
       })
-      const req = makeNavigationEvent("appscheme://connect/loaded")
+      const req = makeNavigationEvent("mx://connect/loaded")
 
       expect(fn(req)).toBe(false)
+    })
+
+    test("it calls loadUrlInBrowser for external urls", () => {
+      const fn = makeRequestInterceptor("https://mx.com/", {
+        onIntercept: (url) => handler(url, {}),
+      })
+      const externalUrl = "https://example.com/page"
+      const req = makeNavigationEvent(externalUrl)
+
+      fn(req)
+      expect(ExpoWebBrowser.openBrowserAsync).toHaveBeenCalledWith(externalUrl)
     })
   })
 
   describe("callbacks", () => {
-    test("calls the onLoadUrlInBrowser callback when loading a valid non post message url", () => {
-      expect.assertions(1)
-
-      const newUrl = "https://mx.com/page2"
-      const callbacks = {
-        onLoadUrlInBrowser: (url: string) => expect(url).toBe(newUrl),
-      }
-
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
-        onIntercept: (url) => handler(url, {}),
-        onLoadUrlInBrowser: (url) => loadUrlInBrowser(url, callbacks),
-      })
-      const req = makeNavigationEvent(newUrl)
-
-      fn(req)
-    })
-
-    test("calls the onLoadUrlInBrowserError callback when loading a url causes an error", () => {
-      expect.assertions(1)
-
-      const newUrl = "https://mx.com/page2"
-      const callbacks = {
-        onLoadUrlInBrowser: (_url: string) => {
-          throw new Error("not today")
-        },
-        onLoadUrlInBrowserError: (url: string, _error: Error) => expect(url).toBe(newUrl),
-      }
-
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
-        onIntercept: (url) => handler(url, {}),
-        onLoadUrlInBrowser: (url) => loadUrlInBrowser(url, callbacks),
-      })
-      const req = makeNavigationEvent(newUrl)
-
-      fn(req)
-    })
-
     test("valid mx/load message calls onLoad callback", () => {
       expect.assertions(1)
 
@@ -105,8 +85,8 @@ describe("makeRequestInterceptor", () => {
       }
 
       const metadata = encodeURIComponent(JSON.stringify({}))
-      const newUrl = `appscheme://load?metadata=${metadata}`
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = `mx://load?metadata=${metadata}`
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -124,8 +104,8 @@ describe("makeRequestInterceptor", () => {
       }
 
       const metadata = encodeURIComponent(JSON.stringify({}))
-      const newUrl = `appscheme://load?metadata=${metadata}`
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = `mx://load?metadata=${metadata}`
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -142,8 +122,8 @@ describe("makeRequestInterceptor", () => {
       }
 
       const metadata = encodeURIComponent(JSON.stringify({ guid }))
-      const newUrl = `appscheme://account/created?metadata=${metadata}`
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = `mx://account/created?metadata=${metadata}`
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -167,8 +147,8 @@ describe("makeRequestInterceptor", () => {
       }
 
       const metadata = encodeURIComponent(JSON.stringify({ user_guid, session_guid, initial_step }))
-      const newUrl = `appscheme://connect/loaded?metadata=${metadata}`
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = `mx://connect/loaded?metadata=${metadata}`
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -188,8 +168,8 @@ describe("makeRequestInterceptor", () => {
       }
 
       const metadata = encodeURIComponent(JSON.stringify({ user_guid, session_guid }))
-      const newUrl = `appscheme://connect/loaded?metadata=${metadata}`
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = `mx://connect/loaded?metadata=${metadata}`
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -204,8 +184,8 @@ describe("makeRequestInterceptor", () => {
         onInvalidMessageError: (url: string, _error: Error) => expect(url).toBe(newUrl),
       }
 
-      const newUrl = "appscheme://connect/notARealMessage?metadata="
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = "mx://connect/notARealMessage?metadata="
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -240,8 +220,8 @@ describe("makeRequestInterceptor", () => {
         }),
       )
 
-      const newUrl = `appscheme://connect/updateCredentials?metadata=${metadata}`
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = `mx://connect/updateCredentials?metadata=${metadata}`
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -258,8 +238,8 @@ describe("makeRequestInterceptor", () => {
       }
 
       const metadata = encodeURIComponent(JSON.stringify({}))
-      const newUrl = `appscheme://load?metadata=${metadata}`
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = `mx://load?metadata=${metadata}`
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)
@@ -275,8 +255,8 @@ describe("makeRequestInterceptor", () => {
         onInvalidMessageError: (url: string, _error: Error) => expect(url).toBeDefined(),
       }
 
-      const newUrl = "appscheme://connect/notARealMessage?metadata="
-      const fn = makeRequestInterceptor("https://mx.com/", "appscheme", {
+      const newUrl = "mx://connect/notARealMessage?metadata="
+      const fn = makeRequestInterceptor("https://mx.com/", {
         onIntercept: (url) => handler(url, callbacks),
       })
       const req = makeNavigationEvent(newUrl)